Disable cache for missing static files
Disable cache with a Cache-Control header for requests made to missing
static file URLs. This is required as Cloudflare will happily cache a
404 page during deployment, which somewhat troublesome.

A better solution obviously would be to upload all the static to a s3
bucket or such before deployment, but this is what we live with for now.