Fix edge case in markdown rendering
Markdown-it-py throws an IndexError with the following content:

> QUOTE\n
+ UNORDERED LIST ITEM\n
  > INDENTED QUOTE\n
\n\n

Removing any of the three content rows or reducing the number of
newlines at the end seems to solve the issue. While stripping the
markdown before rendering might not be the most robust solution, it'll
be enough to fix an issue currently encountered in production.

Refs TS-228