Experimental API: add a permission class based on a shared secret
When completing the OAuth authentication flow, Next.js backend sends a
request to Django backend. To prevent unwanted access to this endpoint,
a shared secret is sent along the request. This was initially included
in the request body and validated in the view. Now the it's included in
the request headers and validated by a custom DRF permission class.

To make the approach reusable, a base class for easily creating
permission classes based on different secrets is provided.

Refs TS-230, TS-354