Make old and new urls work correctly as separate sides of sites (#578)
* Add use_old_urls to CommunityMixin

To be used when both old and new urls are in use at the same time

refs TS-272

* Fix owner_url to work hybrid

* Make old and new urls to work as separate

e.g. links work correctly on each side of the sites

refs TS-272

* Turn error page views to class based

e.g. Handle404 and Handle500
This has to be done, for the base template to work with those views
Also fixup the tests

refs TS-272

* Fix minor bug and improve coverage

When calling templatetag function in tests, the context doesn't contain
the key use_old_urls fix this by doing a get on the dict
Also test old and new url function of the templatetags
Increase coverage

refs TS-272

* Fix breadcrumbs links to work hybrid

Now breadcrumbs will not go cross old and new urls

refs TS-272