Ensure CSRF token is present where needed
Django includes the CSRF token cookie only on pages where the csrf_token
templatetag is used.

As we use it on the API, this means all views which can make API calls
need a CSRF token cookie to be present.

Address the issue by including the CSRF token on all pages which make
API calls.

Refs TS-557