Add format_spec field to PackageVersion
Add a format_spec field to the PackageVersion model, which will be used
to track which format spec a package upload is compliant with.

This will be important when changes are made to the package format, as
downstream tools will have to be able to know if they are able to handle
the packages.

Refs TS-592