Add MirroredS3Storage (#662)
* Add TemporarySpooledCopy context

The context creates a temporary copy of the given file-like, which can
then be operated on without affecting the original one. This comes in
handy e.g. when calling operations that automatically close the file
given as the parameter.

Refs TS-568

* Add MirroredS3Storage

Alternative based on S3Boto3Storage which additionally saves the files
to separately configurable additional S3 mirrors. Mirrors can be used
as backups or when migrating between storage providers.

Refs TS-568