Add API support for legacy profile codes
Add an API endpoint that can be used to create & retrieve legacy
profiles in a similar fashion as they're currently used by r2modman and
Thunderstore Mod Manager.

Previously profile codes have been hosted by Ebkr on a free-tier heroku
instance, and as the free plans are shutting down in the near future,
this API is intended to be used as a replacement.

Refs TS-1116