Speculative fix to upload page error handling bug
A user reported having their entire upload page break when attempting
to upload a package, and the culprit for the breakage seems to be an
undefined value ending in the list of errors when it should only contain
strings.

The root cause was the user's network configuration eating the API
response entirely, which somehow resulted in the above scenario and
broke the page.

This commit is a speculative fix which should hopefully prevent similar
breakage in the future

Refs TS-1479