Filter out deactivated teams from views
Deactivated teams are unusable, yet they're still being shown to the
users as if they were. This commit changes the views to exclude
deactivated views from views where they would normally be visible in.

In practice, this allows for teams to be archived without removing their
members at the time of archival. This could be useful if a team needs to
later be restored and it would otherwise be unclear who is the owner of
the team.

It would be better to have a dedicated field for archival status, but
that's beyond the scope of this PR.

Refs TS-1622