Return user's profile info when logging in via mod manager
Since profile info is "refreshed" when the mod manager is restarted
while storing a valid TS session id, we should store the matching info
when user initally checks in the first time. This doesn't matter too
much now since most of the info is not used on mod manager, but that
can quickly change when new stuff is added to the profile.

Also add a placeholder for user's subscription status to the returned
information.

Refs TS-1640