Improve package list sorting priority
Add upload date as the last sorting criteria for "Most Downloaded" and
"Top Rated" sorting options. This fixes an issue where packages with
equal amounts of ratings or downloads would get arbitrarily sorted in
search results, sometimes leading to a subset of the results never being
visible on any page (if a sufficiently large group of same-priority
sorting packages exists).

Thanks for keks307 on the Inscryption Modding discord for noticing this.

No refs