Improve get_absolute_url of the Package model
Improve the chances of the get_absolute_url function for the Package
model to return an actual valid & functional URL. The function is only
used on Django admin currently, and will generate broken links if the
pacckage exists in a different community than the one bound to the
active domain.

This change makes it so the first active package listing's page URL is
returned instead.