Admin: Improve PackageListing creation performance
The PackageListing creation admin form was getting slowed down due to
listing every category in the form before a community was selected.

Fix this by disallowing selection of categories before the object is
saved once and a community is set, as setting a community will narrow
down the applicable categories to a manageable level.