Add automatic file tree extraction on upload
Automatically extract the file tree of new package uploads as they're
processed.

This can lead to a potential issue with submission processing time
taking much longer than before, and as it's handled synchronously, it
could lead to an increase in submission failures for large packages.

There's no way to fix the above issue without moving to an entirely
asynchronous submission processing, which is a task too large to take on
currently.