Add aggregated fields for Community
Computationally heavy fields, package_count and download_count will be
updated periodically by a background task from now on. They are stored
in a separate model to make it clearer they're not regular fields.

To save memory and improve parallelization, updates are done in steps:
first the related object is created for all Communities that doesn't
have one yet, and then the fields are updated for each Community
separately.

Existing queryset helpers are used to avoid duplicating the logic for
which packages and versions should be included in the calculations.
This means more database operations, but it was deemed a fair tradeoff
since the updates are done as background tasks.

The values should be normally accessed via .aggregated getter to avoid
errors if the Community doesn't have the values computed yet.
.aggregated_fields relation is also available e.g. when working with
QuerySets.

Refs TS-1850