Fix issues in the current user API
Introduce back the old 'teams' field in the current user experimental
API, as it's used by some API clients. Provide the new structure in a
'teams_full' field instead.

Additionally make the OAuth connection rendering not fail if some field
is missing, as we can't guarantee any of the extra data to exist.

Finally take the UserProfileSerializer into use, as it was not actually
doing anything until now.