Make team-related Cyberstorm endpoints support required functionality
Instead of one endpoint for all team information, use separate
endpoints for team's basic information, members, and service accounts.
On the client side these are used in separate tabs, so this works
better with Suspenses, reduces passing props around and avoids
overfetching.

The basic information ("details") is public information, while the
members and service accounts are intended only for team members.

The actual information returned was changed to reflect what's actually
needed on the client.

When checking if the requested team exists, the name is compared case
insensitively. I assume this is the correct approach since creation of
teams with names that differ only in case is prohibited.

Refs TS-1856