Cyberstorm API: add Community's package list endpoint
Add endpoint for fetching package listings related to a specific
community.

The implementation is heavily based on the deprecated
CommunityPackageListApiView under frontend.api.experimental. That
implementation was left as-was for now, cleanup should be done
separately.

The main difference is the inheritance based implemenation which
hopefully allows doing other package list views more easily. Also the
navigation of next/previous pages is done in-line with DRF's
PageNumberPagination which is used in other endpoints of the Cyberstorm
flavour website, whereas the old one used infinite scroll. Also the
page structure is a bit different, so the results don't need to include
e.g. the categories to be used by page's filters.

Refs TS-1875