Cyberstorm API: add Namespace's package list endpoint
Add endpoint for fetching package listings related to a specific
namepace. Since the CommunityListings on backend are bound also to a
Community, the endpoint is also community-scoped. In the future we
might have an endpoint for returning all packages belonging to a
namepace, regardless of the communit, but before that we have to decide
how e.g. category-based filtering would work in such cases.

Refs TS-1875