PackageVersion cursor pagination and db index
Add a function for enumerating over a PackageVersion queryset in a
chunked but deterministic fashion. The implementation follows the cursor
based pagination paradigm where results are sorted in a determenistic
fashion, allowing for consecutive queries to pick off where the previous
left off by using appropriate filtering.

Add a database index which supports the deterministic ordering and
filtering.