Cyberstorm API: refactor package listing endpoints
Original implementation filtered Package's community listings so that
only the listing for current community was included. It was done in a
different part of the code than where the results were serialized. On
serialization it might appear to reader as a bug, since it's not
obvious why the code assumes there to be exactly one community listing
for each package.

To fix this the code that accesses the community is now more explicit.
However, this is done in the base class, which was originally
community-agnostic, but the explicit filtering requires access to
community. To resolve the issue the base class is now community-scoped.

Some QuerySet filters were moved outside the base class to make it more
obvious they don't depend on the class they were in. This change was
also based on code review comments.

Personally I don't know if these changes improve the quality of the
code.

Refs TS-1875