Cyberstorm API: remove caching from package listing endpoints
As per review comments on PR #902, remove the caching from endpoints
that return package listings. The implementation the endpoints had is
not one we want to use in the future. Implementing a generic caching
solution is not something we have time for right now. It was deemed
better to have no caching at all, for now.

Usage of CachedPaginator was the reason the view was very custom
instead of adgering to DRF practices. Now that it's no longer in use,
it's possible to refactor the endpoints to be DRF-ish.

Refs TS-1875