Cyberstorm API: add package details endpoint
Endpoint for fetching the information required to display a package
view. Markdown content is not included since it will be read from a
separate endpoint that processes the raw data to markup in a safe
manner. Also content under different tabs is omitted since they will
be loaded separately when needed.

When the implementation was nearly complete, it occured to us that
we'll actually need a similar endpoint for each versions, whereas the
current implementation assumes the latest version is always requested.
Rather than make this a blocker now, it was decided to move on with the
current implementation and refactor it in the future as needed.

Refs TS-1980