Update dependants link wording
Change the wording used in the dependants link to refer to packages
instead of mods.

This change is done based on a recent discussion in the Lethal Company
community, where it became apparent that using 'Mods' was setting up
users with false expectations, as all kinds of packages are included in
the number (such as modpacks).

While this doesn't solve the use case of "I want to see how many other
mod developers integrate with my mod" (which seemed to be the main use
case), it should at least be a minor improvement over the previous
wording.