Simplify package download view queries
Remove unnecessary queries from the package download view. The extra
queries were in place due to historical reasons which are no longer
valid; specifically the downloads used to be limited to only work in the
domain of the community whose which package is being downloaded. In
practice this filtering has not applied to any new communities since the
transition to subdomainless communities.