Create a new package version download event model
In order to utilize download metrics for time series analysis, the model
structure needs to change drastically. The previous implementation uses
counters instead of storing time-series events, and as such is
conceptually incompatible with a time-series focused data format.

Given the above, the best way forward is to create a new model which
follows a time series paradigm. This commit includes a new
thunderstore.metrics module which should be used for any future
metrics-focused database models.

The new model does not use foreign keys in the schema for future
compatibility reasons, as analytics pipelines commonly have challenges
with supporting constraints like that.