Fix odd issue with test state consistency
Evidently some configurations of test run splitting will lead to a
problem where a test starts off expecting a file to not exist in an S3
bucket, but it does. This problem does not occur if all tests are run.

This does not solve the root cause, but it does address the issue for
the specific test which was failing.