Support mirror CDNs for package and shared profile downloads
API calls can indicate that a mirror CDN is preferred by including a
query parameter in the request. The feature will be supported by the
API endpoints that provide download URLs for packages and shared mod
manager profiles.

The backend will return URL to the preferred CDN only if the CDN is
allowlisted in the config. This is a plain string replacement operation
and the backend doesn't validate the URL it's about to return in any
way.

This is intended to solve the small but growing number of incidents
where users for some reason can't access Thunderstore's main CDN.

Refs TS-2003