Change evaluation of has_changelog on Cyberstorm PackageDetailAPIView
Empty or whitespace-only changelogs are now considered to be valid
changelogs. The rationale is that if the package author includes an
empty changelog file, then we will show an empty changelog. Only if the
file is omitted, will we show "no changelog available".

Also use database level annotation for checking whether the
PackageVersion has a changelog. This is an premature optimization based
on an assumption that DRF's method serializers can be slow.

Refs TS-2006