Cyberstorm API: unify team views and URLs
- Drop "Detail" from resource's GET endpoint.
- Use "List" for endpoints that return lists.
- Drop trailing "s" from nouns in the URLs

Ideally adding a new team member would have been done by posting to
´team/member/´ rather than a separate ´team/member/add´ endpoint.
However, it seems DRF is not particularly good at combining different
methods in one class, so I left it as was for now.

This is part of an effort to make the names of the views and the
related URLs logical, descriptive, and consistent.

Refs TS-2008