Cyberstorm API: include deactived dependencies for package listing
Even if a dependency is deactived, it's still a dependency of the
package, which might not work without it. To avoid misleading
information on the website, include deactived dependencies too.

PackageVersion is considered deactived if the parent Package is
deactived, even if the specific version would be active.

Annotating the "effective active status" to the QuerySet was
considered, but since the related package is prefetched for other
purposes anyway, it seemed like an unnecessary step.

Refs TS-2013