Cyberstorm API: censor deactived package dependencies
The reason why a package has been deactivated may be an obnoxious icon
or description text, so don't send those to clients.

Method serializers are used instead of annotations, since especially
for the icon field the annotation isn't straightforward: the ImageField
can't be directly attached to the object, and parsing the absolute URL
correctly from the path representation stored in the database is
error-prone.

Refs TS-2013