Bolster zip file validation
Include validation that ensures the zip file starts at offset 0 and that
it includes no relative paths in the files.

In theory the zip files might still contain bogus data in the middle,
but this is fairly unlikely to happen in practice and shouldn't impact
the behavior of other applications.