Account for tests in multi-cache support
Add a setting which can be used to disable the use of multiple caches,
leading to the default cache being used for all purposes.

This is mainly intended to be used for the test runner as the test
fixtures aren't configured with multiple cache dbs in mind.

Additionally, use a proxy object for accessing the caches in order to be
compatible with what Django does by default and thus what the cache test
fixtures have been configured to handle.