Remove legacy download logging, use celery
Move the download count logging to an asynchronous backround task as
keeping it blocking the request is silly.

This should be moved off of Celery entirely later and instead e.g. to
Kafka, but as Celery is already available in the environment, it's a
good starting point.

This commit also removes the legacy download logging which has not been
in use in any environment for a while.