Create VisibilityFlags + add to PackageVersion
Create a new VisibilityFlags model which is intended to hold object
visibility level information. Also include a VisibilityMixin for ease of
implementation to models as well as VisibilityQuerySet to make querying
with correct visibility settings easier.

Take the VisibilityMixin into use on the PackageVersion model. Query
changes have to wait the visibility table to be populated first.