Add 'installers' field support to package manifest
Add support for a new 'installers' field to the package manifest.json.
This is intended to allow packages to declare which install strategies
a mod manager should apply to the package when installing.

This commit does not know or care what the installers actually do or how
the installer system works, it only focuses on validating that the
declared installers are known to exist in the ecosystem. What ultimately
populates the "known installers" db table is left as a consideration for
later.