Support mirror CDNs for chunked package list index download
Calls to the API endpoint can indicate that a mirror CDN is preferred
by including a query parameter in the request. The endpoint will return
an URL to the preferred CDN only if the CDN is allowlisted in the
config.

This feature is intended to mitigate the issues of the CDN provider.