Exclude package listings with shared communities
Add method to PackageListingQueryset class that filters PackageListings
to include only listing associated with a single community.

The method uses annotate to count the distinct communities associated
with each package and filthers the results to include only those with
a count of 1.

Add the filter to the listings queryset in the update_for_community
method. This will ensure that only package listings associated with
a single community is counted in the aggregated fields.

Refs. TS-2265