Implement is_removed
Implement a is_removed property in the Package model that determines
whether a package is removed by checking if is_active is set to False
or if the Package has no active versions.

Implement a is_removed property in PackageVersion that determines
whether a package version is removed by 1) checking if the related
Package is removed 2) checks if is_active is set to False.

Refs. TS-2275