Refactor filter_by_review_status
Update the function to take community as an argument.

Update the query to user filter instead of the double negative exclude
filtering. This query seemed to not exclude packages with a rejected
listing if the packages were multi-community packages. It seems to not
handle rejected listings explicitly for a given community, which returns
unwanted results.

Refs. TS-2346