Update user delete view queries
Change the get queries in DeleteUserAPIView and
DisconnectUserLinkedAccountAPIView to not use iexact since this might
retrieve wrong user.

Update DisconnectUserLinkedAccountAPIView with a helper function
to check whether the provider is actually the last one instead of just
checking the count. This will prevent the view from returning a 403
instead of returning a 404 when a 404 is expected.

Update tests to use response.json() in all assertions.

Refs. TS-2317 & TS-2318