Fix Manual Download btn styling
Update styling for the "Manual Download" button to use a conditional
check. If the listing can be installed with mod manager, and the
"Install with mod manager" button is rendered, the styles remain the
same as before. However, if the "Install with mod manager" button is not
rendered, the "Manual Download" button will adjust to the same size as the
"Support the creator" button for a more consistent layout.

Refs. TS-2367