Attempt to improve chunked package list API response performance
The older enpoint uses django.utils.cache.get_conditional_response.
Cloudflare uses some of the headers the function checks but testing
this locally in a manner that would match reality wasn't feasible.

Last-Modified header was added to a normal response since the older
endpoint does so, although it's unlikely to have an effect with the
redirect response this endpoint is returning.

Although the changes are speculative, they were low effort to
implement. Whether they have effect requires a deployment to the
production environment.

Refs TS-2376