Update community caches in parallel
Refactor task for updating chunked community packages to update each
community in parallel. Introduce a new task which is spawned by the main
update task to fire on each community.

Remove function responsible for updating all community caches and
introduce a new function doing the same thing but for a single
community, taking the community as a parameter.

Update cache tests because the old cache update function was removed.

Add update_single_community_cache to celery tests.

Add a create_batch method to CommunityFactory to help with testing.

Add test for the newly updated celery task to assert that the child
tasks are indeed spawned per community.

Refs. TS-2378