Pre release fixes
Update the URL for creating new from "/new" -> "/create". Update tests
accordingly.

Fetch team with only name instead of name__iexact in several Team views.

Update DisbandTeamAPIView to use APIView instead of DestroyAPIView. Add
a function called disband() to Team model which takes care of the actual
removing of the Team, as well as checks the permissions. This change in
order to move the deletion logic out of the view and in to the model
where it can be reused.

Update tests accordingly.

Refs. None