Refactor team disband and create views
Use the team services create_team and disband_team in the form views
responsible for creating and disbanding teams. Catch the errors manually
which are raised in the services and add them to the form.

Remove redunant functionality from the forms, such as validations that
are handled in the services and remove the save functions from the forms
since this is handled by the services.

Refs. TS-2426