Extend exception_handler with new exceptions
Implement a custom exception to be used with the service layer
functions. Most of the functions raise ValidationErrors where
PermissionDenied errors should be raised in the API views. Catch these
exceptions in the custom exception handler and re-raise the errors as
PermissionDenied exceptions.

Add functionality for catching Django's built-in PermissionError as
PermissionDenied errors in the custom exception handler since this error
is used in at least two places: reject and approve package listings.

Refs. TS-2429