Inherit django validation error in custom error
Update the PermissionValidationError to inherit ValidationError from
django.core.exceptions in order to ensure backwards compatibility.

Update the exception_handler.py to check for the
PermissionValidationError first, followed by the other checks with elif
statements. We need to do this in order to seprate bewteen
PermissionValidationError and ValidationError so we do not mutate the
"exc" variable with the wrong type of error.

Refs. TS-2429