Update service layer exception handling
Rely on implicit exception handling rather than re-mapping exceptions to
simplify code.

An argument could be made that the control flow is cleaner with explicit
try/catch logic as exceptions are bad for control flow and it's not
immediately clear the functions will throw (and there's no way to
statically check for this), but as a similar convention has already been
used elsewhere in the codebase, it seems better than duplicating the
validation logic implementation.

An alternative would be to implement validation in a way which doesn't
directly raise an exception but instead returns a list of errors as a
return value, which could be used in an if/else check on the call site
to provide a more explicit control flow. Refactoring validation to work
in this manner is beyond the scope of this pull request however, so
that's left as a potential future area of improvement.