User service layer in team update form view
Use the service layer function in the form responsible for updating
team's donation_links. Utilize the service layer function in the form's
save function and update the view accordingly.

In order to invoke the post-validation errors added to the form in the
form tests, call form.save() in the tests as well.

Implement tests for the view in a similar manner as they are implemented
for the forms to assert functionality.

Refs. TS-2314