Refactor disband_team and create_team
Refactor and rename the parameters in the service layer functions for
creating and disbanding teams. Pass a Team object to the disband_team
function, the service layer function should not be responsible for
throwing a 404 if the team object is not found, this is the
responsibility of the view(s).

Refactor tests accordingly to the changes made in the functions.

Refs. TS-2426