Use service layer in form views
Use the service layer functions for disbanding and creating teams in the
form views responsible for the same actions.

Update the form's save() functions to use the service layer functions.
Update the tests accordinly to these changes. Some of the errors are now
added to the form in the form's save function, which means the save()
function has to be called in the tests in order to invalidate the form.

Refs. TS-2426