Use service layer functions in form views
Utilize create_team and disband_team in form views. Update forms to use
the service layer functions in the form's save() functions and update
the view logic accordingly.

Update the form for creating teams to be a Form instead of ModelForm in
order to simplify the validations since validation happens in the
service layer function and also in the model.

Update all the tests accordingly to these changes. Add more tests to the
service layer functions.

Refs. TS-2426