Rework markdown service and celery task
Rewrite the polling logic in the service to use the get() function from
celery instead to wait for the result of a task. Raise errors when
expected, the .get() function raises either a TimeoutError or Exception.

No need to pass the statys_key to the celery task anymore due to the
changes made in the service.

Update tests accordingly.

Refs. TS-2516