Fix potential state bug in DragDropFileInput
Assigning `fileInputRef.current` to a const at the start of the
component is invalid, as it will contain stale data in the event the ref
gets updated later on. As the event handlers were relying on the const
instead of the fileInputRef.current, they would have accessed stale
data rather than the updated version.

In practice this did not cause issues in the current version of the code
as the ref isn't updated, but it's good to address either way.