Implement package dependencies list api view
Implement a view responsible for fetching dependencies for a package.

The view returns a list of paginated and serialized package versions
related to a specific package.

Pagination is set to 20 items per page.

The view expects a version number, or alternatively "latest" in the URL
kwargs to be present. When using "latest", the version is set to None
which is handled by get_package_version().

Refs. TS-2639