Update form views and forms
Update the forms and form views for removing user and removing user's
social accounts.

Implement tests.

Update template to displaty error. A user should not be able to remove
the last social auth account but it can be bypassed by removing the
disabled attribute with deveoper tools.

Refs. TS-2317 & TS-2318