Use services in create/disband form views
Use the create_team and disband_team services in the forms and views
responsible for disband teams and creating teams. Update tests
accordingly.

Update the form for creating teams to clean the name manually in order
to keep the error messaging consistent, otherwise tests will fail since
because it is a ModelForm which throws a different formatted error by
default.

Refs. TS-2426