Disallow "latest" in dependencies list URL
Remove the possibility to use the "latest" keyword in the dependencies
list URL for fetching dependencies for a package version. The previous
implementation did not require a specific version number and allowed for
using "latest" as a keyword for getting the dependencies for latest
version. Remove this feature and force usage of specific version number
in the URL in order to improve caching possibilities.

Refs. TS-2731