Simplify KafkaClient and kafka configuration
Remove the KAFKA_DEV setting and replace it with KAFKA_TOPIC_PREFIX, as
that is the only actually relevant piece of configuration that should be
exposed; the clients behave the exact same way outside of which prefix
is being used.

This change also leads to simplification of the `ts_analytics/kafka.py`
module code.