Move Kafka config validation to happen at startup
Move the Kafka config validation step to take place on Django startup
(when settings are loaded) rather than the moment it's actually being
used, as that way configuration errors become immediately apparent
(e.g. when deploying an update to them) rather than giving the
impression the configuration is valid and then later on failing anyway.