Delete unused `close` function from KafkaClient
The KafkaClient's `close` function is never used and rather misleading
(it does in fact not close the producer, simply calls flush on it), not
to mention there's no good handling of a potential timeout (or a way
to configure it).

As such it makes sense to delete the entire problematic snippet of code,
which this commit does.