Cyberstorm API: add endpoint for package listing version
Conceptually this muddies the water since listing is package scoped.
However, the UI supports showing the community listing page for the
latest version (default), or a specific version. We want to use this
endpoint since it verifies that the listing is not rejected, ensuring
we don't serve malicious content on the web page. This is in contrast
to PackageVersionAPIView which doesn't check rejection status, as it
doesn't have a way to so since it doesn't have access to community.

In the future we could consider having the package listing endpoint
which returns only the listing and package specific fields, and a
separate endpoint which returns the package version related fields
while checking that the parent package is actually available. This
would remove the conceptual ambiguity and improve cacheability.